% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/food_dat.R
\docType{data}
\name{food_dat}
\alias{food_dat}
\title{Food Data Set}
\format{
An object of class \code{data.frame} with 387 rows and 20 columns.
}
\source{
Lotspeich SC, Mullan AE, McGowan LD, Hepler SA. Combining
Straight-Line and Map-Based Distances to Investigate the Connection Between
Proximity to Healthy Foods and Disease. Stat Med. 2025 Mar 30;44(7):e70054.
doi: 10.1002/sim.70054. PMID: 40226886; PMCID: PMC11995689.
}
\usage{
food_dat
}
\description{
Merged Analytical Dataset created using the Healthy Food Store Locations,
Disease Prevalence by Neighborhood (Census Tract), Neighborhood
(Census Tract) Population Centers, and Rural-Urban Commuting Area (RUCA)
Codes.
}
\keyword{datasets}
