% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conc_curve.R
\name{make_conc_curve}
\alias{make_conc_curve}
\title{Plot a concentration curve
This function plots a concentration curve using \code{ggplot2}}
\usage{
make_conc_curve(conc_data)
}
\arguments{
\item{conc_data}{pre-processed dataset of ordered, aggregate data returned by \code{make_conc_data} function}
}
\value{
A \code{ggplot} object of the concentration curve.
}
\description{
Plot a concentration curve
This function plots a concentration curve using \code{ggplot2}
}
