% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_conc_data.R
\name{make_conc_data}
\alias{make_conc_data}
\title{Construct dataset to build a concentration curve
This function creates an ordered, aggregate (where applicable) dataset from which to build a concentration curve}
\usage{
make_conc_data(
  data,
  outcome_var,
  rank_var,
  rank_ascend = TRUE,
  round_rank = NULL,
  group_var = NULL,
  na.rm = TRUE
)
}
\arguments{
\item{data}{dataset containing at least \code{outcome_var} and \code{rank_var}}

\item{outcome_var}{variable name of the outcome, to be passed through to \code{dplyr} (so no quotation marks)}

\item{rank_var}{variable name of the disadvantage or socioeconomic exposure used to rank, to be passed through to \code{dplyr} (so no quotation marks)}

\item{rank_ascend}{logical, if \code{TRUE} (the default) then rows are ranked from smallest to largest values of \code{rank_var}}

\item{round_rank}{(optional) numeric, use to round \code{rank_var} before ordering and grouping. If \code{NULL} (the default), no rounding is done.}

\item{group_var}{(optional) variable name of a grouping variables, to be passed through to \code{dplyr} (so no quotation marks)}

\item{na.rm}{logical, if \code{TRUE} (the default), then rows of \code{data} where \code{outcome_var} and/or \code{rank_var} is missing are removed}
}
\value{
A \code{data.frame} is returned:
\item{delta}{Estimated average treatmentf effect on the primary outcome.}
}
\description{
Construct dataset to build a concentration curve
This function creates an ordered, aggregate (where applicable) dataset from which to build a concentration curve
}
